Created WaitCondition for cluster
Now we are waiting for all hadoop-ec2 instances signals' before
proceeding with creation of Administrator server instance. Which means
that they should be online and done with the bootstrapping before
starting work on the Administrator server. This is achieved by signaling
for each hadoop-ec2 instance with a single wait handle with multiple
count based on cluster size. Then configured a DependsOn property on
Administrator server against the waitcondition.